+++ src/turn/frontend.spec.ts
@@ -1,7 +1,7 @@
 /// <reference path='../../../third_party/typings/es6-promise/es6-promise.d.ts' />
 /// <reference path='../../../third_party/typings/jasmine/jasmine.d.ts' />
 
-import turn_frontend = require('../turn-frontend/turn-frontend');
+import turn_frontend = require('./frontend');
 import messages = require('./messages');
 import net = require('../net/net.types');
